pyproject.toml
[project]
name = "duckdb-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "duckdb>=1.3.1",
    "gdown>=5.2.0",
    "office365-rest-python-client>=2.6.2",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.1",
]
